+++ package.json
@@ -17,7 +17,7 @@
     "fs:read": "node src/fs/read.js",
     "fs:rename": "node src/fs/rename.js",
     "hash": "node src/hash/calcHash.js",
-    "modules": "node src/modules/esm.mjs",
+    "modules": "node --no-warnings src/modules/esm.mjs",
     "streams:read": "node src/streams/read.js",
     "streams:transform": "node src/streams/transform.js",
     "streams:write": "node src/streams/write.js",
